feat: add copy and deepcopy to base
